IsMounted: factored out getDevice()
Currently getDevice is not needed, but it might
be in the future, so factor it out.

I also experimented a bit to make IsMounted() a bit faster,
tried using substring match instead of a regexp. Practically
it makes no difference as the overhead of parsing a string is
negligible compared to that of fork/exec.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>