SetVerbosity(): fully implement
We need to fiddle with various situations wrt verbosity,
such as then -vN is explicitly specified, when output to
stdout is needed by ploopOut, etc. This patchset fixes
all such issues to the best of my knowledge, i.e. now
this package works with any verbosity levels (except
for an issue in Mount() that is addressed by the next
commit).

Fix README.md accordingly.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>