tests: run with minimal output by default
As we can set verbosity now, let's run tests with minimal verbosity,
only showing the errors. If required, modify prepare() and change
the argument of SetVerbosity() to whatever you like.

Note benchmarks are run with NoStdout verbosity anyway, as otherwise
the output is all cluttered and it's hard to see results.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>