Allow to create a ploop volume in a pre-created directory
Some vstorage attributes (e.g. encoding parameters) has to be set on a
directory, and they are applied only for new files.

So if a user wants to set encoding parameters, he has to create a
directory, set vstorage attributes on it and only then he can create a
ploop volume.

Signed-off-by: Andrei Vagin <avagin@openvz.org>